+++ R/030_init.R
@@ -29,6 +29,8 @@ class_train.kknn <- new_S3_class("train.kknn")
 class_bartmodel <- new_S3_class("bartmodel")
 class_hal9001 <- new_S3_class("hal9001")
 class_earth <- new_S3_class("earth")
+class_lmtree <- new_S3_class("lmtree")
+class_glmtree <- new_S3_class("glmtree")
 
 
 # %% --- Generics -------------------------------------------------------------------------------------